Report undefined symbols when only producing a wasm file
When building direclty to wasm (e.g. with `-o <out>.wasm`
or with `--oformat=wasm`) we don't run the JS compiler
after wasm-ld.   This means that we were failing to report
undefined symbols at all (since we pass `--allow-undefined`
to wasm-ld).

In this mode enable LLD_REPORT_UNDEFINED which has pretty
much the exact behaviour we need.  It run the JS compiler
pre-link to produce a list of valid undefined symbols.

Fixed: #12725